agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "visitcare-agent"
version = "0.1.0"
description = "訪問介護AIエージェント（シフト管理AI + ヘルパー支援AI）"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.28",
    "google-cloud-firestore>=2.19",
    "firebase-admin>=6.6",
    "google-cloud-aiplatform>=1.70",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "gunicorn>=23.0",
    "httpx>=0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
lint.select = ["E", "F", "I", "W"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
